Fix missing chars in altium-ibom-generator
The current (direct) HTML generator needs embedded (SVG) font data to draw the texts. It currently lacks a few chars, and thus kicks them out of the texts.
If we're using the InteractiveHtmlBom executable, we can let it handle the texts with all chars in them, and use its default font (newstroke).